projects/ig-scraper-brander/requirements.txt
# Instagram Video Scraper & Brand Overlay Tool
# Install with: pip install -r requirements.txt

# Core Instagram scraping
instaloader>=4.10.0

# Video processing
ffmpeg-python>=0.2.0
opencv-python>=4.9.0

# Web app
flask>=3.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
Pillow>=10.0.0

# Optional: For better video handling
moviepy>=1.0.3
